python-project/shared_core/requirements-dev.txt
pydantic
python-dotenv
oracledb
pydantic-settings
aiokafka